+++ ReviewNow/Controllers/UsersController.cs
@@ -1,21 +1,28 @@
 ﻿using Application;
+using AutoMapper;
 using Domain;
+using Domain.NormalDomain;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using ReviewNow.ExportDtoClases;
 using System;
+using System.Threading.Tasks;
+
 namespace ReviewNow.Controllers
 {
     [ApiController]
     [Route("[controller]")]
     public class UsersController : ControllerBase
     {
-        private readonly ILogger<UsersController> logger;
-        private readonly IUsersRepostory _userRepository;
+        private readonly ILogger<UsersController> _logger;
+        private readonly IUsersRepository _userRepository;
+        private readonly IMapper _mapper;
 
-        public UsersController(ILogger<UsersController> logger, IUsersRepostory userRepository)
+        public UsersController(ILogger<UsersController> logger, IUsersRepository userRepository, IMapper mapper)
         {
-            this.logger = logger;
-            this._userRepository = userRepository;
+            _logger = logger;
+            _userRepository = userRepository;
+            _mapper = mapper;
         }
 
         [HttpGet]
@@ -25,17 +32,23 @@ namespace ReviewNow.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(User user)
+        public async Task<IActionResult> CreateAsync([FromBody] UserDto userDto)
         {
-            _userRepository.Add(user);
-            return Ok("I did it!!!!!");
+            User userToAdd = _mapper.Map<User>(userDto);
+            User user=await _userRepository.AddAsync(userToAdd);
+            UserExpDto userExpDto = _mapper.Map<UserExpDto>(user);
+            return Created("~",userExpDto);
         }
 
-        [HttpDelete]
+        [HttpDelete("{userId}")]
         public IActionResult Delete(Guid userId)
         {
+            if (_userRepository.Find(userId) == false)
+                return NotFound();
+
             _userRepository.Delete(userId);
-            return Ok("I removed it!!");
+
+            return NoContent();
         }
     }
 }
